Allow Extension URLs to load unsandboxed
This may sound like a really bad change to make, however I have some reasons as to why I believe this is perfectly fine:
1. Why is it fine to load text unsandboxed?
There's no difference between a URL extension loading unsandboxed and a text extension using eval to load all of it's code from another website.
2. Sharing & updating convenience
If I made an extension and already had it hosting perfectly fine, I would rather share the link to that hosting than tell everyone to download it and keep coming back to the website incase there is an update. Again, malicious updates would still be possible if a file-loaded extension just used eval & fetch to run all of it's code.